% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roo.R
\name{conformal.pred.roo.surv}
\alias{conformal.pred.roo.surv}
\title{In-sample split conformal prediction intervals for right-censored data.}
\usage{
conformal.pred.roo.surv(
  x,
  t,
  d,
  tau,
  train.fun,
  predict.fun,
  w = NULL,
  cens.model = "km",
  alpha = 0.1,
  mad.train.fun = NULL,
  mad.predict.fun = NULL,
  split = NULL,
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{Matrix of features, of dimension (say) n x p.}

\item{t}{Vector of responses (observed times), of length (say) n.}

\item{d}{Vector of responses (censoring indicator), of length (say) n.}

\item{tau}{Horizon of time.}

\item{train.fun}{A function to perform model training, i.e., to produce an
estimator of E(min(T*,tau)|X), the conditional expectation of the true 
event time T* given features X. Its input arguments should be x: matrix of 
features, t: vector of observed times, d: vector of censoring indicators
and tau: horizon of time. train.fun can be a function performing model 
training for several (say m) regression models.}

\item{predict.fun}{A function to perform prediction for the (mean of the)
responses at new feature values. Its input arguments should be out: output
produced by train.fun, and newx: feature values at which we want to make
predictions.}

\item{w}{Censoring weights. This should be a vector of length n. Default is 
NULL, in which case censoring weights are computed using the censoring 
model specified by the parameter cens.model, or by default with the 
Kaplan-Meier model.}

\item{cens.model}{Model used to estimate the censoring survival function.
This should be equal to "km", "rsf" or "cox" (respectively Kaplan-Meier, 
Random Survival Forests or Cox). Default is "km".}

\item{alpha}{Miscoverage level for the prediction intervals, i.e., intervals
with coverage 1-alpha are formed. One or several values can be passed for
which to compute intervals. Default for alpha is 0.1.}

\item{mad.train.fun}{A function to perform training on the observed absolute 
residuals R = |min(T,tau) - m(X)| and their associated censoring weights,  
where m denotes the estimator produced by train.fun, to estimate E(R*|X)
for the true absolute residual denoted R*. 
This is used to scale the conformal score, to produce a prediction interval
with varying local width. The input arguments to mad.train.fun should be
x: matrix of features, y: vector of observed absolute residuals, and w: 
censoring weights. The default for mad.train.fun is NULL, which means that 
no training is done on the absolute residuals, and the usual (unscaled) 
conformal score is used. Note that if mad.train.fun is non-NULL, then so 
must be mad.predict.fun (see next).}

\item{mad.predict.fun}{A function to perform prediction for the (mean of the)
absolute residuals at new feature values. Its input arguments should be
out: output produced by mad.train.fun, and newx: feature values at which we
want to make predictions. The default for mad.predict.fun is NULL, which
means that no local scaling is done for the conformal score, i.e., the
usual (unscaled) conformal score is used.}

\item{split}{Indices that define the data-split to be used (i.e., the indices
define the first half of the data-split). Default is NULL, in which case
the split is chosen randomly.}

\item{seed}{Integer to be passed to set.seed before defining the random 
data-split to be used. Default is NULL, which effectively sets no seed.
If both split and seed are passed, the former takes priority and the latter
is ignored.}

\item{verbose}{Should intermediate progress be printed out? Default is FALSE.}
}
\value{
A list with the following components: pred, lo, up, fit, split,
  fit.all, out.all. pred and fit are matrices of dimension n x m. lo and up 
  are arrays of dimension n x m x a. Recall that n0 is the number of 
  rows of x0, m is the number of regression models and a is the number of 
  values passed to alpha. Hence, the rows of the matrices pred, lo, up give 
  the predicted value, and lower and upper confidence limits (from rank-one-
  out split conformal inference), respectively, for the response at the n 
  points given in x. The rows of the matrix fit give the fitted values for 
  the n points given in x. The indices used for the half of the data-split 
  are returned in split. Finally, for convenience, the output from running
  train.out on the entire (unsplit) data set x,y is stored in out.all, and
  fit.all is an n x m matrix whose rows store the fitted values from out.all.
}
\description{
Compute prediction intervals for for restricted time-to-events based on right-
censored data, having asymptotically valid in-sample coverage, using rank-
one-out (ROO) split conformal inference.
}
\references{
"Distribution-Free Predictive Inference for Regression" by Lei,
  G'Sell, Rinaldo, Tibshirani, Wasserman (2018). See also "A Comprehensive 
  Framework for Evaluating Time to Event Predictions using the Restricted 
  Mean Survival Time" by Cwiling, Perduca, Bouaziz (2023) for the extension
  to right-censored data.
}
\seealso{
\code{\link{conformal.pred}}, 
  \code{\link{conformal.pred.roo}}, \code{\link{conformal.pred.split.surv}}
}
