% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformalInference.R
\docType{package}
\name{conformalInference-package}
\alias{conformalInference}
\alias{conformalInference-package}
\title{Tools for conformal inference in regression}
\description{
A collection of tools for distribution-free inference for regression problems
  using the theory of conformal prediction.
}
\details{
Conformal inference is a framework for converting any pre-chosen estimator of
  the regression function into prediction intervals with finite-sample
  validity, under essentially no assumptions on the data-generating process
  (aside from the the assumption of i.i.d. observations). The main functions
  in this package for computing such prediction intervals are
  \code{\link{conformal.pred}} (shorter intervals, but more expensive) and
  \code{\link{conformal.pred.split}} (wider intervals, but much more
  efficient).
}
