% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{interval.build}
\alias{interval.build}
\title{Helper function to join each B prediction intervals in Multi Split
For each point in x0, it combines the simulated B intervals into one.}
\usage{
interval.build(yyy, B, tr)
}
\arguments{
\item{yyy}{column vector of B lower bounds and B upper bounds}

\item{B}{number of replications}

\item{tr}{truncation threshold for the algorithm}
}
\description{
Helper function to join each B prediction intervals in Multi Split
For each point in x0, it combines the simulated B intervals into one.
}
