% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw.R
\name{ipcw}
\alias{ipcw}
\title{Computation of Inverse Probability Censoring Weights. The censoring survival 
function can either be estimated with Kaplan-Meier, Cox or Random Survival 
Forests.}
\usage{
ipcw(t, d, x, tau, cens.model = "km")
}
\arguments{
\item{t}{Vector of responses (observed times), of length (say) n.}

\item{d}{Vector of responses (censoring indicator), of length (say) n.}

\item{x}{Matrix of variables, of dimension (say) n x p.}

\item{tau}{Horizon of time.}

\item{cens.model}{Model used to estimate the censoring survival function.
This should be equal to "km", "rsf" or "cox" (respectively Kaplan-Meier, 
Random Survival Forests or Cox). Default is "km".}
}
\value{
A vector of length n containing the censoring weights.
}
\description{
Computation of Inverse Probability Censoring Weights. The censoring survival 
function can either be estimated with Kaplan-Meier, Cox or Random Survival 
Forests.
}
