% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loco.R
\name{loco.surv}
\alias{loco.surv}
\title{Variable importance via median excess test error extended to right-censored
data}
\usage{
loco.surv(
  x,
  t,
  d,
  tau,
  train.fun,
  predict.fun,
  w = NULL,
  active.fun = NULL,
  vars = 0,
  alpha = 0.1,
  rho = 0.5,
  bonf.correct = FALSE,
  split = NULL,
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{Matrix of features, of dimension (say) n x p.}

\item{t}{Vector of responses (observed times), of length (say) n.}

\item{d}{Vector of responses (censoring indicator), of length (say) n.}

\item{tau}{Horizon of time.}

\item{train.fun}{A function to perform model training, i.e., to produce an
estimator of E(min(T*,tau)|X), the conditional expectation of the true 
event time T* given features X. Its input arguments should be x: matrix of 
features, t: vector of observed times, d: vector of censoring indicators
and tau: horizon of time. train.fun can be a function performing model 
training for several (say m) regression models.}

\item{predict.fun}{A function to perform prediction for the (mean of the)
responses at new feature values. Its input arguments should be out: output
produced by train.fun, and newx: feature values at which we want to make
predictions.}

\item{w}{Censoring weights. This should be a vector of length n. 
Default is NULL, in which case censoring weights are computed using the 
Kaplan-Meier model.}

\item{active.fun}{A function which takes the output of train.fun, and reports
which features are active for each fitted model contained in this output.
Its only input argument should be out: output produced by train.fun.}

\item{vars}{A list specifying the variables (indices between 1 and p)
for which variable importance should be investigated. Alternatively, if
set equal to 0, the default, then all variables are investigated. If both 
vars and active.fun are passed, the latter takes priority and the former
is ignored.}

\item{alpha}{Miscoverage level for the confidence intervals, i.e., intervals
with coverage 1-alpha are formed. Default for alpha is 0.1.}

\item{bonf.correct}{Should a Bonferroni correction be applied to the p-values
and confidence intervals? Default is FALSE.}

\item{split}{Indices that define the data-split to be used (i.e., the indices
define the first half of the data-split, on which the model is trained).
Default is NULL, in which case the split is chosen randomly.}

\item{seed}{Integer to be passed to set.seed before defining the random 
data-split to be used. Default is NULL, which effectively sets no seed.
If both split and seed are passed, the former takes priority and the latter
is ignored.}

\item{verbose}{Should intermediate progress be printed out? Default is FALSE.}
}
\value{
A list with the following components: inf.sign, active, master, 
  bonf.correct. The first is a list, containing the results of LOCO inference
  with the sign text adapted to right-censored data. This list has one 
  element per regression model the training and prediction functions, 
  train.fun and predict.fun. The inherent to second returned component active
  is a list, with one element per tuning step, that reports which features 
  are active (or at least investigated) in  the corresponding fitted model.
  The third returned component master collects all active (or investigated) 
  features across all regression models, for easy reference. The last
  component signals whether a Bonferroni correction has been applied.
}
\description{
Compute confidence intervals for median excess test error due to 
  dropping a variable. Extension to estimators of the restricted mean
  survival time with right-censored data.
}
\details{
The sign test adapted to right-censored data is asymptotically valid
  if the Kaplan-Meier model is used to estimate the censoring survival 
  function in the censoring weights, under the assumption of independent 
  censoring.
}
\references{
"Distribution-Free Predictive Inference for Regression" by Lei,
  G'Sell, Rinaldo, Tibshirani, Wasserman (2018). See also "A Comprehensive 
  Framework for Evaluating Time to Event Predictions using the Restricted 
  Mean Survival Time" by Cwiling, Perduca, Bouaziz (2023) for the extension
  to right-censored data.
}
