% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.pred.R
\name{plot.rmst.pred}
\alias{plot.rmst.pred}
\title{Plot function for rmst.pred object.}
\usage{
\method{plot}{rmst.pred}(x, elements = c("all"), model.names = NULL, varsL = 0, ...)
}
\arguments{
\item{x}{The rmst.pred object.}

\item{elements}{Selection among "mse", "vimpL", "vimpG" or "all"describing 
which elements of the analysis to show. If "mse" then the results on the 
estimation of the mean squared error are displayed; if "vimpL" then results
of local variable importance are displayed; if "vimpG" then results of 
global variable importance are displayed. If "all" then the results from 
all elements are displayed. Default is "all".}

\item{model.names}{Names given to the different learning models given to
train.fun. Default is NULL.}

\item{varsL}{A list specifying the variables (indices between 1 and p)
for which local variable importance should be plotted. Alternatively, if
set equal to 0, the default, then local variable importance is plotted for
all variables.}

\item{...}{Other arguments (currently not used).}
}
\description{
Plot function for rmst.pred object.
}
