% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loco.R
\name{print.loco}
\alias{print.loco}
\title{Print function for loco object.}
\usage{
\method{print}{loco}(x, test = c("wilcox", "sign", "z", "all"), digits = 3, ...)
}
\arguments{
\item{x}{The loco object.}

\item{test}{One of "z", "sign", "wilcox", or "all", describing which test 
results to show. If "z", then the results from only Z-tests are displayed;
if "sign", then only sign tests; if "wilcox", then only Wilcoxon tests. If
"all", then the results from all tests are displayed. Default is "wilcox".}

\item{digits}{Number of digits to display. Default is 3.}

\item{...}{Other arguments (currently not used).}
}
\description{
Summary and print the results of a set of simulations, stored an object
of class sim (produced by sim.master).
}
