% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loco.R
\name{print.loco.surv}
\alias{print.loco.surv}
\title{Print function for loco.surv object.}
\usage{
\method{print}{loco.surv}(x, digits = 3, ...)
}
\arguments{
\item{x}{The loco.surv object.}

\item{digits}{Number of digits to display. Default is 3.}

\item{...}{Other arguments (currently not used).}
}
\description{
Print function for loco.surv object.
}
