% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.pred.R
\name{print.rmst.pred}
\alias{print.rmst.pred}
\title{Print function for rmst.pred object.}
\usage{
\method{print}{rmst.pred}(
  x,
  elements = c("all"),
  digits = 3,
  model.names = NULL,
  var.names = NULL,
  ...
)
}
\arguments{
\item{x}{The rmst.pred object.}

\item{elements}{Selection among "mse", "range", "vimp", or "all", describing
which elements of the analysis to show. If "mse", then the results on the 
estimation of the mean squared error are displayed; if "range", then
information on the range of the prediction intervals are displayed; if 
"vimp", then results of global variable importance are displayed. If
"all", then the results from all elements are displayed. Default is "all".}

\item{digits}{Number of digits to display. Default is 3.}

\item{model.names}{Names given to the different learning models given to
train.fun. Default is NULL.}

\item{var.names}{Names given to the different variables used in the 
regression. Default is NULL.}

\item{...}{Other arguments (currently not used).}
}
\description{
Print function for rmst.pred object.
}
