% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{print.sim}
\alias{print.sim}
\title{Print function for sim object.}
\usage{
\method{print}{sim}(x, type = c("err", "len", "all"), se = TRUE,
  digits = 3, ...)
}
\arguments{
\item{x}{The sim object.}

\item{type}{One of "all", "err", or "len" describing the type of summary to 
be used. If "err", then for each method, the tuning parameter value that
gives the best average test error is selected, and the coverage and length
properties at this tuning parameter value are report. If "len", then the
analogous thing is done, but for the tuning parameter value that gives the
best average length. If "all", then the average coverage, length, and test
error results are shown across all tuning parameter values. Default is
"err".}

\item{se}{Should standard errors be displayed (in parantheses)? For each
metric, this is defined as the standard deviation divided by the square
root of the number of simulation repetitions. Default is TRUE.}

\item{digits}{Number of digits to display. Default is 3.}

\item{...}{Other arguments (currently not used).}
}
\description{
Summary and print the results of a set of simulations, stored an object
of class sim (produced by sim.master).
}

