% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{print.tex}
\alias{print.tex}
\title{Print function for latex-style tables.}
\usage{
\method{print}{tex}(tab, tab.se = NULL, digits = 3, file = NULL,
  align = "l")
}
\description{
Print a given table in format digestable by latex.
}

