% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf.funs}
\alias{rf.funs}
\title{Random forests training and prediction functions.}
\usage{
rf.funs(ntree = 500, varfrac = 0.333, replace = TRUE,
  obsfrac = ifelse(replace, 1, 0.632), nodesize = 5)
}
\arguments{
\item{ntree}{Number of trees to grow. Default is 500.}

\item{varfrac}{Determines the number of variables used as candidates at each
split: this is floor(varfrac * p), where p is the total number of
variables. Default is 0.333.}

\item{replace}{Should sampling of observations done with or without
replacement? Default is FALSE.}

\item{obsfrac}{Determines the number of observations used by each tree: this
is floor(obsfrac * n), where n is the number of training observations.
Default is 1 when replace is TRUE, and 0.632 when replace is FALSE.}

\item{nodesize}{Sought size of terminal nodes, in growing each tree. Default
is 5.}
}
\value{
A list with two components: train.fun and predict.fun.
}
\description{
Construct training and prediction functions for random forests.
}
\details{
Based on the package \code{\link{randomForest}}. If this package is
  not installed, then the function will abort.
}
\examples{
## See examples for sam.funs function
}
\author{
Ryan Tibshirani, friends
}
\references{
\url{http://www.stat.cmu.edu}
}

