% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.xy}
\alias{sim.xy}
\title{Features and responses generation.}
\usage{
sim.xy(
  n,
  p,
  x.dist = c("normal", "binom", "sn", "mix"),
  cor = c("none", "pair", "auto", "rand"),
  rho = 0.5,
  k = 5,
  standardize = TRUE,
  mean.fun = c("linear", "additive", "rotated"),
  m = 4,
  sparsity = c("strong", "weak"),
  s = round(log(p)),
  weak.decay = 0.5,
  mu.seed = NULL,
  error.dist = c("normal", "unif", "t", "sn"),
  sigma = 1,
  bval = 3,
  snr = NULL,
  sigma.type = c("const", "var"),
  df = 3,
  alpha = 5,
  omitted.vars = FALSE
)
}
\arguments{
\item{n, p}{The dimensions, as in n x p, of the feature matrix x. Note that n
is also the length of the response vector y.}

\item{x.dist}{One of "normal", "binom", "sn", "mix", indicating the
distribution for the entries of x. In the first three cases, the columns
of x are filled out with i.i.d. normal, standard binomial, or skewed normal
(with skewness parameter alpha = 5) draws, respectively. The third option
requires the \code{\link{sn}} package. In the last case, "mix", each column
of x is drawn with equal probability from the above three distributions.
Default is "normal".}

\item{cor}{One of "none", "pair", "auto", "rand", indicating the type of
correlation among columns of x. If "pair", then x is post-multiplied by
the symmetric square root of the p x p matrix whose diagonals are 1 and
off diagonals are rho (see parameter below). If "auto", then each column of
x is taken to be a random convex combination of its old value and the
previous k (see parameter) columns. If "rand", then each column of x is
taken to be a random convex combination of its old value and a randomly
k other columns. Default is "none".}

\item{rho}{If cor is "pair", then this specifies the pairwise correlation.
Default is 0.5.}

\item{k}{If cor is "auto" or "rand", then this specifies the number of other
columns that are used in the lagged or random convex combinations (see
above). Default is 5.}

\item{standardize}{Should the columns of x be standardized (i.e., centered,
and scaled to have variance 1)? Default is TRUE.}

\item{mean.fun}{One of "linear", "additive", or "rotated", indicating the
functional dependence of the mean on the variables. If "additive", then the
mean vector mu is constructed to be an additive function of B-splines of
the variables, where each variable is expanded into df B-splines (with
knots spaced evenly over the range of observed measurements of the
variable). If "rotated", then mu is set to be randomly rotated version of
such an additive function. Default is "linear".}

\item{m}{Number of B-splines per variable in the "additive" or "rotated" 
cases for mean.fun; default is 4. Ignored when mean.fun is "linear".}

\item{sparsity}{Either "strong" or weak". In the former case, s (see below)
randomly chosen variables of the p total contribute to mu, with
coefficients randomly set to +1 or -1. In the latter case, still s variable
contribute to mu, as before, but now the remaining p-s variables contribute
in a weak sense, in that their coefficients are randomly set to +1 or -1
times a successively increasing power of the weak.decay factor (see below).
In the "additive" or "rotated" cases for mean.fun, the above should be
interpreted in the appropriate groupwise sense, e.g., s groups of spline
functions are randomly chosen, each group corresponding to the expansion of
a single variable.}

\item{s}{Number indicating the level of sparsity. Note that the choice s = p
would correspond to a fully dense model. Default is round(log(p)).}

\item{weak.decay}{Number controlling the decay of the weak coefficients, in
the case sparsity is "weak", as their magnitudes get set to a successive
power of weak.decay. Default is 0.5.}

\item{mu.seed}{Seed to be set before generation of the mean vector mu
(ensures that the mean is the same, over multiple calls to this function).
Default is NULL, which effectively sets no seed.}

\item{error.dist}{One of "normal", "unif", "t", "sn", indicating the error
distribution, where the last option "sn" is for the skewed normal, and
requires the \code{\link{sn}} package. Default is "normal".}

\item{sigma}{Standard deviation of errors. When the error distribution does
not have a second moment (i.e., t-distribution with 1 or 2 df), then sigma
represents the MAD. Default is 1.}

\item{bval}{Magnitude of nonzero coefficients in the model for mu as a linear
function of x (or as a linear function of B-splines of x in the "additive"
and "rotated" cases for mean.fun). Default is 3. This is overriden if the
snr parameter is specified (see next).}

\item{snr}{The signal-to-noise ratio, specifically var(mu)/sigma^2, where
sigma^2 is the variance of the errors (see previous). The mean will be
scaled to meet the desired signal-to-noise ratio.}

\item{sigma.type}{Either "const" or "var", indicating constant or varying
error variance. In the latter case, the error standard deviation sigma is
multiplied by 1 + abs(mu)/mean(abs(mu)) (this gives a vector of the same
length as mu, i.e., a different standard deviation per component).}

\item{df}{If error.dist is "t", then this specifies the degrees of freedom of
the t-distribution for the errors. Default is 3.}

\item{alpha}{If error.dist is "sn", then this specifies the skewness
parameter (i.e., shape parameter) of the skewed normal distribution for the
errors. The location and scale parameters are set so that the resulting
distribution always has mean zero and variance 1. Note that alpha > 0 means
right-skewed and alpha < 0 means left-skewed. Default is 5.}

\item{omitted.vars}{Should important (active) variables be omitted from x?
Induces correlation between the x and the error terms.}
}
\value{
A list with the following components: x, y, and mu.
}
\description{
Generate a feature matrix x, and response vector y, following a specified
setup.
}
\details{
In the case of an "additive" or "rotated" mean, this function relies
  on the packages \code{\link{plyr}} and \code{\link{splines}}. If these
  packages are not installed, then the function will abort.
}
\examples{
## See examples for sam.funs function
}
\references{
"Distribution-Free Predictive Inference for Regression" by Lei,
  G'Sell, Rinaldo, Tibshirani, Wasserman (2018).
}
