% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.pred.R
\name{wrss}
\alias{wrss}
\title{Mean Squared Error of a Restricted Mean Survival Time estimator}
\usage{
wrss(
  x,
  t,
  d,
  tau,
  train.fun,
  predict.fun,
  w = NULL,
  cens.model = "km",
  CV = T,
  n.folds = 10,
  split = NULL,
  p = 0.1,
  seed = NULL
)
}
\arguments{
\item{x}{Matrix of variables, of dimension (say) n x p.}

\item{t}{Vector of responses (observed times), of length (say) n.}

\item{d}{Vector of responses (censoring indicator), of length (say) n.}

\item{tau}{Horizon of time.}

\item{train.fun}{A function to perform model training, i.e., to produce an
estimator of E(min(T*,tau)|X), the conditional expectation of the true 
event time T* given features X. Its input arguments should be x: matrix of 
features, t: vector of observed times, d: vector of censoring indicators
and tau: horizon of time. train.fun can be a function performing model 
training for several (say m) learning models.}

\item{predict.fun}{A function to perform prediction for the (mean of the)
responses at new feature values. Its input arguments should be out: output
produced by train.fun, and newx: feature values at which we want to make
predictions.}

\item{w}{Censoring weights. This should be a vector of length n. Default is 
NULL, in which case censoring weights are computed using the censoring 
model specified by the parameter cens.model, or by default with the 
Kaplan-Meier model.}

\item{cens.model}{Model used to estimate the censoring survival function.
This should be equal to "km", "rsf" or "cox" (respectively Kaplan-Meier, 
Random Survival Forests or Cox). Default is "km".}

\item{CV}{Boolean indicating whether or not to perform cross-validation.
Default is TRUE.}

\item{n.folds}{The number of folds for the cross-validated estimation of the
mean squared error. The default number is 10 folds.}

\item{split}{Indices that define the data-split to be used (i.e., the indices
define the training set) in case cross-validation is not performed. 
Default is NULL, in which case the split is chosen randomly.}

\item{p}{Proportion of data to put in the test set in case cross-validation 
is not performed. If p=0 then the whole data set is used both as training 
set and test set.}

\item{seed}{Integer to be passed to set.seed. Default is NULL, which 
effectively sets no seed.}
}
\value{
A matrix of dimension n.folds (1 if cross-validation is not 
  performed) x m containing the estimation of the mean squared error using 
  censoring weights on each fold and for each learning model.
}
\description{
Estimation of the mean squared error of a learning algorithm for the 
restricted mean survival time trained on right-censored data.
}
\details{
The estimation of the mean squared error with the WRSS estimator
can be performed in one of the following ways: 

- By cross-validation on a defined number of folds;
- By dividing the data into a training set and a test set, on which to train
the learning algorithm and compute the estimator of the mean squared error,
respectively;
- By using all data both as training set and test set.

The default method is the first one.
}
\references{
See "A Comprehensive Framework for Evaluating Time to 
  Event Predictions using the Restricted Mean Survival Time" by Cwiling, 
  Perduca, Bouaziz (2023).
}
